Changed the starting weapons and weapon behaviors in player.def, gauntlet def, and blaster.def. Hopefully the hitscan charge shot works. I couldn't test it since the game is giving me such a weird buf that's preventing me from playing.
The bug is prompting none of the shortcut commands to work, and is prompting me to enter a CD key that I never needed. I assume this was because I extracted all the game files from the game folders for easy access of all game files.That must have triggered something in the game to respond to a possible piracy threat. Now I need to find out how to solve this, Otherwise my work will be down the drain :(