10/31/18 - Final commit for Project Submission
Finalized changes.
Added readme with detailed instructions for how to test the mod changes and a list of all completed deliverables.
Ready for submission and grading.
---------------------------------------------------------------
The first half of this semester has been absolutely insane. But I'm happy to say that I made it through. Now here's to hoping the rest of it goes well.
I want to thank my professors and my friends for getting me through this rough life patch. I don't think I would have been here long enough to see the end of this project had it not been for their support. Thank you, from the bottom of my heart, truly.